regler les imports dans les scripts de web scraping, correction du dockerfile pour le lancement du workernode celery
