+++ Data_extraction/Websites/MarocAnn.py
@@ -9,7 +9,7 @@ from selenium.common.exceptions import (
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
-from selenium_init import (
+from Websites import (
     check_duplicate,
     init_driver,
     load_json,
