+++ Data_extraction/Websites/Rekrute.py
@@ -5,7 +5,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
-from selenium_init import (
+from Websites import (
     check_duplicate,
     init_driver,
     load_json,
