+++ Data_extraction/Websites/bayt.py
@@ -14,7 +14,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
-from selenium_init import (
+from Websites import (
     check_duplicate,
     init_driver,
     load_json,
@@ -236,7 +236,7 @@ def change_page(
         return False
 
 
-def main():
+def bayt_scrape():
     start_time = time.time()
     logger.info("Début de l'extraction des offres d'emploi sur Bayt.com")
     # Initialiser le driver
@@ -271,4 +271,5 @@ def main():
     return data
 
 
-main()
+if __name__ == "__main__":
+    bayt_scrape()
