--- celery/test_celery.py
@@ -1,3 +1,0 @@
-from tasks import add
-
-add.delay(2, 5)
