+++ celery_app/tasks.py
@@ -1,0 +1,6 @@
+from celery_app import app
+
+
+@app.task
+def add(x, y):
+    return x + y
